NOT YET WORKING: modification of syscall exit path
see comments in arch/x86/entry/entry_64.S for specifics
This implementation works in simple tests i.e. elevating, reading CR3,
and lowering. Also tested a normal, unelevated getpid loop and it does
not disrupt functionality for lowered threads making syscalls.
Mitigations do not function, but using the interposing mitigator the
double fault mitigation can be deployed, but text fault mitigation
fails. The error seems to occur after an elevated thread is interrupted.
In gdb, observed that on the return from interrupt path the registers
struct passed to exit_to_usermode_prepare, and irqentry_exit_to_usermode
is broken. The values are there but off by 1 i.e. rax is in ip, ip is in
cs, cs is in flags, flags is in sp etc. think it might have something to
do with how IRET behaves wrp which registers it pops depending on
whether it believes a change in privilege is happening. Need to look
into it further.